+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/formatting/ResendingTextFormatting.kt
@@ -2,7 +2,7 @@ package dev.inmo.tgbotapi.extensions.utils.formatting
 
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.message.textsources.TextSourcesList
-import dev.inmo.tgbotapi.types.ParseMode.*
+import dev.inmo.tgbotapi.types.message.*
 import dev.inmo.tgbotapi.types.message.content.TextContent
 
 fun createFormattedText(